view tests, tests into separate files
